Rework sprites and positioning, use size 64x64, add themes and more
Now the cell size is 64x64 instead of 50x50.

Now there is a themes concept. Add theme "classic" with existing images.

Now "bones", "crack" and "rocks" are transparent and are applied to "floor".

Now the the sprite positioning is at center and not at top-left.

Now the code is much more independent of cell sizes.

Changed values of number of enemies and hero health, these are
auto-calculated now.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>