Improve arrow key behaviour and make diagonal moves configurable
Now if you release a key that is already processed (i.e. animated the
move to right) then the next move in that direction is not done.

Now it is possible to press left and then right (or up and then down),
and it will process these 2 keys in the pressed order on the separate
iterations (previously the move was voided, since it was -1 and then +1).

Now the iteration step is 0.18 second, this is also the move animate
duration.

Currently diagonal moves are enabled. Will be decided later.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>