Add the concept of level time limit, use limit of 1 minute in level 3
Now there is a level time shown in the status area.

When a level has a time limit, the count down is shown instead, and it
becomes red when critical time is remaining.

Update the defeat condition to this case when the level has a time limit
and it is reached.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>