Implement color puzzle level concept and add new puzzle level
The target of color puzzle is to press plates and convert colors to a
single color. Pressing a plate advances a color of all neighbour cells.

Make color puzzles configurable with N colors (2 to 6) and with size
SIZE_XxSIZE_Y where size is an odd number.

Plates are generated in (x, y) possitions where x and y are odd.

To press a plate, player must stay on it and press Space key.

Add "plate" image and "floor_gray_alpha" image needed for color puzzle
levels (themes).

Add level 4 that is a color puzzle 11x11 with 6 colors and timeout.

Use a new music tone in the new level.

Refactor some code and fix some bugs in generate_level (f.e. the floor
additions were not regenerated). Add new victory condition.

Change playable grid from 12x10 to 11x11, so that it is sutable for color
puzzles too.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>