Add support for new level features (gate_puzzle and portal_end)
The level's portal_end flag changes the level target from
kill-all-enemies to reach-the-end-portal. It places the portal at the far
most accessible cell in the level.

The level's gate_puzzle flags triggers gate puzzle with plates and gates.
Each plate in this puzzle triggers one or more gates. The goal is to reach
the end portal; i.e. the gate_puzzle flag assumes portal_end.

Currently the gate puzzles are created without checking that they are
solvable.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>