Support reload of the same level layout by pressing Alt-r
When pressing "r", the whole level is regenerated. But sometimes there is
a need to reload the level with the original generated puzzle, barrels,
gates, char position and enemies.

So, store all of this in stored_level and implement reload_stored flag in
init_new_level function, activated by Alt-r.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>