Implement spiral_maze and grid_maze for levels
In the "spiral maze", you start in the top left and finish in the center.
A spiral maze may be either the one when you first go right and the one
when you first go down; this is chosen randomly.

In the "grid maze", the cells with even x and even y are borders, other
cells are floor.

Add a concept of any_maze that is either random_maze, spiral_maze or
grid_maze.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>