Rework cell types including floor and portal, add start and finish types
What was called portal now is called finish. But stoneage_puzzle has
both. Add support for "start" and "finish" cell type in themes; these are
start and end some puzzles and the level can request these using "start"
and "finish" flags in config.

Minor refactoring of some cell type code and renamings in it to make the
code more readable.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>