Refactor functions working on actor to new CellActor class
Create new CellActor with public methods working with pos and
inplace_animation.

Support internal values for scale (0 .. 1), angle (any angle in degrees,
like 0 .. 720), opacity (0 .. 1) and flip (bool for X and bool for Y).

The inplace animation is activated by giving one or more ranges of:
scale, angle, opacity, flip, and animation duration.

Such animated inplace actors are automatically added to
active_inplace_animation_actors and removed from it when the duration is
passed. The main update function automatically updates the animation of
all such configured actors.

Currently the animation for START and FINISH cells is activated (the char
appears from nothing with a rotation, and disappears to the perspective,
when applicable).

Also, now enemies are killed using new inplace animation mechanism, they
fall either to the right or to the left and then disappear.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>