Implement portal concept, redesign external actor animate, fix keys
Portals are created in pairs. Each portal remembers the desctination of
the the pair.

Move animate of actors into CellActor class, it gets the same arguments,
but may be simplified, since the end position is now by default the
proper actor destination that is usually desired. Also, it now defines
own on_finished so the actor knows when it is animated externally. The
user's on_finished is supported too when given; in inplace animation too.

The portal animation is to disappear in the source, then to teleport and
apper at the destination. This is implemented using external animation of
move with on_finished, then using inplace animation with on_finished.

Now arrow keys are disabled when during either inplace or external
animation of the char. This fixes some bugs.

Also, add 2 new BEAT config constants instead of hardcoding 0.3 seconds.
And add a new beat-enemy animation for falling.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>