Implement lock puzzle as a new level mode
The lock puzzle consists of keys and locks. It also includes the finish
cell. It is somewhat similar to the game puzzle, but instead of pressing
plates to open gates, the corresponding keys should be gathered to open
locks.

There are 2 kinds of keys and locks, just like in the stoneage puzzle.
These are called key1 and key2 (and lock1 and lock2 correspondingly).

The number of locks is configurable by "min_locks" and "max_locks", the
default is 2 to 4 locks per room.

The number of gathered keys of each type is now displayed in the status.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>