Add enemy-dropping-key feature to lock puzzle, enemies containing keys
Now there is a variation of lock puzzle that is controlled by boolean
level flag "enemy_key_bonus". Instead of keys on the map, enemies contain
keys of the needed type that are dropped when a player kills that enemy.

Implement this using enemy bonuses. This implementation is currently in
parallel to the regular mode of the lock puzzles. It is planned that both
variations will use the bonus functionality; this may be implemented
after the planned bonus code refactoring soon.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>