Refactor all bonus code into new class Drop, simplify and clean the code
Now everything that was called "bonus" is removed. Instead, a new class
Drop defines the whole life flow of the drops that may fall from enemy.

The new class Drop represents the drop type and has methods:

* contain(enemy)     - call on enemy creation
* instantiate(enemy) - call on enemy kill, or drop is placed on the map
* collect(cell)      - call on collecting drop from the map
* consume            - call when drop is used

The first stage is optional. When the drop is placed on the map directly,
call instantiate with cell and not with actor.

The last stage is optional. Intended for keys that are consumed at
corresponding locks.

There are currently 4 drop types (objects): heart, sword, key1, key2.

Now, the keys in lock and stoneage puzzles converted to the new drop
interface (there are 2 modes, with enemy holding the key, and without).

So, map cell types CELL_KEY1 and CELL_KEY2 are removed.

Removed all bonus lists; now the drop object contains the list of cells
with this type of drops on the map.

Added new heart and sword images, they are bigger and made 64x64.

The status drop images are scaled 0.7 from the regular drop images.

There is a generic code to draw all drops on the map and in the status.
The status drops are drawn centered; only the active drops are drawn.
The drop is considered active if there are instances of it kept in the
enemy or on the map.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>