Implement stopless mode when char is moved until end without stopping
This mode is triggered by pressing RShift + S.

Currently no puzzle uses this mode, but future puzzles may use it.

In this mode, now the move is divided into 2 phases in some cases, the
first is free move (like with lifts, just on a floor), and the second is
to complete the single move (beat enemy, move barrel or collect a drop).

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>