Add support for nine_rooms for to all puzzles
So now it is possible to divide the map to 9 rooms 3x3 (for the default
play size) with the same puzzle in each room, just like it is possible to
divide to 4 rooms 5x5 (for the default play size).

The room parameters for nine_rooms are constant, just like it was with
four_rooms, and now flags.py handles both cases, when to use room-4
parameters and when room-9 parameters.

Update main and color/lock puzzles accordingly.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>