Add completely new MemoryPuzzle with multiple options
The memory puzzle is to find matching pairs of numbers (or colors).

The puzzle is defined by "memory_puzzle": True in levels.

The puzzle size is given by "memory_puzzle_size", similarly to
"color_puzzle_size". However only even number of cells participates. When
the odd dimentions are given, then the central cell remains empty, and it
is used as starting char cell.

When the number of pairs is 6 or less (i.e. the puzzle size is 12 or
less), then 6 colors (the same as defined for ColorPuzzle) are used
instead of numbers.

The pairs are open by pressing Space. If two open cells contains the same
number or color, then they disappear. The puzzle goal is to open all
matching pairs.

If the second open cell does not match, then both open cells are closed.
There is a configurable time when the open cells are closed (1 second).
When the Space is pressed they may be closed earlier.

It is possible to configure reveal_time per level to reveal the numbers
or colors at the start to help a player. There is a configurable
reveal-fade-time (1.5 seconds) during this time the revealed numbers or
colors fade out. When space is pressed, then this initial reveal is
canceled.

It is possible to always reveal cells. In this case, the puzzle is to
find the visible pairs without a need to memorize them, usually with
time_limit.

The character is drawn with opacity (configurable, by default 0.4) on
closed or open number cells. When the colors are used, this is not used.

The provided 5 levels demonstrate different puzzle modes, four_rooms,
nine_rooms, with or without reveal, with has_start and has_finish,
with or without time_limit, with or without monsters.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>