Add completely new FifteenPuzzle with configurable size
The famous fifteen puzzle by LLoyd. The target is to solve the scrambled
puzzle with numbers and an empty cell, so that the numbers are sorted and
the empty cell is at lower right position.

It is possible to move a cell into the empty cell by pressing Space. Then
the current cell becomes empty.

Four levels from different sizes and with different numbers of rooms
added, with sizes 3x2, 3x3, 4x4, 7x7 and different options like the
finish cell and enemies.

See also: http://migo.sixbit.org/puzzles/fifteen/

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>