Support cropping the bg image and dividing a large image into cells
It is possible to configure now "bg_image_crop" in levels. In this case
the image will first be maximally cropped while retaining the proportions
(ratio) and then scaled (with the same ratio in this case) as usual.

Add create_cell_subimage to be able to divide the large image into cells
and even rotate it (mostly useful with angles 0, 90, 180, 270).

Also, apply_diff and CellActor apply and move methods now support
"subtract" or "undo" flag that applies the opposite diff.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>