Add completely new RotatepicPuzzle that involves rotating picture boxes
This is a classic puzzle that divides the picture into cells, scrambles
them by rotating into 90 factor angles and presents to the user to rotate
then back to the original picture.

These new level fields are supported:
* rotatepic_puzzle
  - boolean to turn this puzzle on
* rotatepic_puzzle_size
  - size in cells like (5, 5); the default is the whole room
* rotatepic_puzzle_image
  - the large image that is divided into boxes, like pictures/my.png
* rotatepic_puzzle_image_crop
  - flag to maximally crop the picture retaining its proportions

To rotate the box, go to it and press Space.

To rescramble, press Backspace. To show the solved position, press Enter.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>