Improve the room concept and generalize Puzzle set_area_from_config
Now when there are no multi-rooms, the room idx is 0 and not None,
NUM_ROOMS is 1 and not None, and there is new MULTI_ROOMS flag.

This simplifies most of the code, since there is always one room now or
more (four_rooms or nine_rooms).

Many puzzles need to have puzzle-specific area (that is just a part of
the playing room). Generalize it to Puzzle.set_area_from_config with
options min_size=None, request_odd_size=False, align_to_center=False.

Change the Puzzle subclasses to use new method and remove duplicate code.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>