Support loading map from file and add original stoneage level #1
Now when "map_file" key is specified in the level config, its value is
the map file to be loaded instead of generating map and rooms.

The map file consists of the first (signature) line that looks like
"# Dungeon map 12x10" that encodes the map size as the last word.

Then there are map chars (similar to debug_map() output) with one char
(cell type) per cell. The lift characters are converted to actors and the
map cell is set to CELL_VOID, other cells are set just from the file.

The portals are currently not supported; they require the
portal_destination metadata that will be added later.

The new load_map_file function tries to check all failures in the map
file and just ignores the map file with the printed error if something is
wrong (like invalid or unexpected map size in the file or unexisting file
or invalid format).

We are now ready to add the first original stoneage level to demonstrate
this new feature. More stoneage levels will be added later.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>