Support puzzles that use the whole map and do not need border of walls
Usually, the map [0, N-1] is constructed using a border of walls that
includes play area [1, N-2]. When the play area is divided to 4 or 9
rooms, this is done by adding inner borders to the outer border.

Support a new border-less mode when the map [0, N-1] does not initially
have any border of walls and the play area is [0, N-1] too. When the map
is divided to 4 or 9 rooms, this is done by inner borders only, and the
rooms are slightly bigger (since 2 cells are added instead of border).

Now the order of operations on level initialization is this:
    flags.parse_level(level)
    puzzle = create_puzzle(level)
    set_map_size(level.get("map_size", DEFAULT_MAP_SIZE), puzzle.has_border())
    flags.apply_sizes()

Change the code to allow new dependencies. This is done by breaking
flags' calls into two. After the first parse_level call the number of
rooms is available, but not the room sizes that will be available only
after the second apply_sizes call.

For example, stoneage (and new atomix puzzle, to be added soon) are
lift-based (lifts move on void, not on floor) and do not need walls
surrounding the level map.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>