Multiple small fixes and preparations for inclusion of new AtomixPuzzle
1) convert_inner_walls now can get optional cell_type to be used instead
of CELL_INNER_WALL. I.e. it is possible to turn it into VOID or FLOOR.

2) Change assert_room to allow access to room (of the previous level)
during mode=next just like with mode=game and mode=init.

3) Add function get_actors_in_room.

4) If there are no empty cell for initial char placement, but there are
lifts, allow to place char on the first lift.

5) In create_lift allow to pass custom surface to be used for lift image
instead of theme image "lift<X>".

6) Add function get_lift_target_at_neigh.

7) Export more functions to puzzle using Globals:
	is_cell_in_map
	is_cell_in_room
	get_actors_in_room
	get_lift_target_at_neigh

8) Fix theme switching when lift has custom surface (see item #5). In
this case, the custom surface is preserved and no theme image is loaded.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>