Add cursor concept with selected actor that is not just char
So now all atomix puzzles are solvable.

The Cursor is kind of CellActor with added functionality:
	reset()
	is_active()
	is_char_selected()
	is_lift_selected()
	toggle()
	selected_actor (either char or lift or cursor itself)

When Enter is pressed without active cursor then the cursor is activated.
When the cursor is active it is possible to move it through the room
freely. When the active cursor is pressed, then the actor on that cell is
selected if char or lift, otherwise the cursor is deactivated.

The selected char may be moved just like as it always worked.
The selected lift may be moved independently of the char now.

Pressing Escape or Space resets cursor (i.e. selects the char for
movement).

Added cursor image that is created by us and is freely distributable.
Later the cursor image may become theme-able.

Currently cursor is always active-able. Later this feature may be enabled
for certain levels only.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>