Allow to beat or kill enemy even through the barrel
When pushing a barrel into an enemy, reuse the same code as in the direct
enemy confrontation, just the char health is not harmed. I.e. allow to
beat or kill an enemy staying on the barrel way.

Added char's (and barrel's) animation on beat similarly to the enemy beat
animation, but with a different tween. The beat animation for char (and
barrel if applicable) is activated even when the enemy gets killed.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>