Add solver to BarrelPuzzle and solution mode to solve level automatically
Implement DFS algorythm of finding solution or barrel (sokoban) puzzles.

It is currently not optimal, but works well for small levels.

The solution mode is activated by KP Enter key and disactivated by any
other key.

In the solution mode, the solution moves are taken every 0.3 seconds.
There is support for beating and killing enemies. I.e. the same move may
be taken multiple times until the char gets to the desired cell.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>