Fix reloading the same level (on Alt-R) for different puzzles
in FifteenPuzzle, GatePuzzle, MemoryPuzzle, PortalPuzzle and
RotatepicPuzzle implement the store_level and restore_level interface to
keep the internal data.

Also, to make all puzzles that use self.map (for example, BarrelPuzzle)
to be restorable on Alt-R, call on_create_map when reloading the same
level after restoring the map.

Also, to make LockPuzzle and StoneagePuzzle to be restorable on Alt-R,
enhance Drop objects with new restore and store interface and call it on
reloading the same level.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>