Add convert_outer_floors to be used with loaded barrel maps
In barrel puzzle, we generate the random puzzle with initially all walls,
so in on_generate_map() we want to call convert_inner_walls().

And the map files are all floor instead, so in on_load_map() we want to
call convert_outer_floors().

The effect is the same: the outer (inaccessible) area is converted either
to CELL_VOID (if bg_image is given) or to CELL_OUTER_WALL.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>