Introduce enter_room() and puzzle.on_enter_room()
Previously set_room() was used on both generation and playing.
This is still the case, but separate playing from generation by
adding additional enter_room() and puzzle callback on_enter_room.

Now puzzles may do play-only initialization when starting to play a room
(entering the room) that are not called on generation of room as well.

This also simplified the code and fixed some things.

Now char_cells (was: char_cell) can be saved for each room upon room
generation (or on Alt-R) and not just for the first room.

The set_char_cell() interface is not changed by default, but it now sets
the char_cell for the currently set room unless the new optional explicit
parameter of room_idx is given. This new parameter is used in load_map,
since by definition load_map is called when there is no room set (it
itself sets single room on load success).

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>