Add new script tools/convert-mosaic-map
Mosaic is our different project with the mode to easily edit dungeon or
even stoneage maps.

This new script gets either a mosaic save file as parameter (and then it
converts it to dungeon .map format) or a dungeon map file as parameter
(and then it converts it to mosaic .sav format).

Usage:
	tools/convert-mosaic-map ../mosaic/stoneage-1.sav \
		"Original StoneAge level 1" >maps/stoneage/1.map
or:
	tools/convert-mosaic-map stoneage-1.map >../mosaic/stoneage-1.sav

The special filename "-" is supported too, then the input map or sav is
taken from stdin.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>