Select char and not lift by cursor when character stays on lift
Previously when cursor selects a cell with the char staying on a lift,
the lift is selected, so char can't move into a different direction from
the lift. This is not desired. This was not very problematic, since it
was still possible to unselect the lift by pressing Escape, but let's
just select the char in this case.

We still support the selected lift when the char stays on it (just in
case this may be needed in the future), but after this change it is
impossible to come to this situation when the lift under the char is
selected.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>