Add Puzzle press_cell interface invoked by cursor, support it in 4 puzzles
When Enter is pressed, the cursor is activated. And when the cursor is
active already, Enter deactivated it. Now, when the cursor is active, a
new Puzzle press_cell hook is called. The puzzle may process it.
Moreover, the puzzle may signal not to deactivate the cursor by returning
True in press_cell. If it returns False, the usual cursor deactivating
happens.

The more correct hook name would be on_enter_pressed_when_cursor_active,
but press_cell is an intuitive name too.

Support this new press_cell interface in color, fifteen, memory and
rotatepic puzzles. This is the same as coming to the cell with char and
pressing Space. Just remotely, no need to move character.

This makes the enemies useless in these puzzles, since it is possible to
press cells even with enemies on them or on the way to them. But it's OK,
since these puzzles are not enemy oriented.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>