Auto-activate cursor when the selected one-direction lift was moved
Implement the behaviour of the original StoneAge. When the one-direction
lift is moved, return the cursor to the active state, instead of keeping
the lift active when it can't move any more.

Wait for the lift animation to be finished before activating the cursor.

It is still possible for now to select the lift that can't move, but this
should be done universally (even for bi-directional lifts) by checking
all possible moves.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>