Add support for mouse in puzzles using press_cell
Extend the puzzle.press_cell interface to have button parameter.
By default it is None (when cell is pressed using keyboard Enter).

Add support for 5 or more mouse buttons: left (1), middle (2), right (3)
and wheel's up (4) and down (5).

In ColorPuzzle, it does not currently matter which mouse button is
pressed, the effect is the same as with keyboard - rotate colors up.

In FifteenPuzzle and MemoryPuzzle, only the left button is handled, the
same as with keyboard. All other mouse buttons are ignored.

In RotatepicPuzzle, handle all mouse button, i.e. enhance the puzzle to
be able to rotate clockwise (on keyboard Enter with cursor, on keyboard
Space without cursor, on right mouse button and wheel down), to rotate
counterclockwise (on left mouse button and wheel up) and to rotate
upsidedown (on middle mouse button). Note that previously it was rotated
counterclockwise with keyboard, but in the code it was called clockwise;
make it behave as it was intended in the code.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>