Enhance and slightly fix check_victory conditions and messages
Now check_victory informs the player about the goal.

When the level target is to solve puzzle then it now writes either
"Solve puzzle!" or "Puzzle solved!".

When the level target is to kill all enemies then it now writes either
"Kill all enemies!" or "All enemies killed!".

When the level has finish, it now writes "Reach finish!" if needed.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>