Only show cursor coordinates when debug is on, and improve it
Fix canceling cursor using Escape, now this clears the cursor coordinates
as intended, just as with canceling it using the second Enter.

Use set_status_message2 to show cursor coordinates, not to change the
main status_message, although this is not really needed, since this main
status message is constantly regenerated.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>