Implement Puzzle.parse_config_num method
This method gets puzzle config key (like "num_plates") and the default
value. It supports number or tuple or range as config value. In the last
two cases a random number from the tuple or the range is choosen. Because
of this feature, this method should be called only once per config key,
and the return value should be stored in the puzzle object if needed.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>