GatePuzzle: Fix Solution.open_gates calculation
Only combined_gate_bits were stored in SpanModel, so it was not really
possible to accurately calculate the gates that switched the state from
closed to open. Store both combined_gate_bits and toggled_gate_bits per
plate, to use both for different needs.

Now Solution.open_gates accurately indicates the gates that ever became
open, even when the combined plates were pressed in one span, and one
plate opened the gate and another plate closed the same gate.

There is still a minor situation when 3 span plates close, then open,
then close again the same gate. Then open_gates will not be set for that
gate. But this is not very problematic, since this 3 (or 5, or 7) actions
of the neerby same-span plates may be thought as not opening that gate.
The situation of open, then close, then open again is detected correctly.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>