GatePuzzle: Allow 4 gates to be attached to a plate and require all gates
There were [1..3] gates attached to one plate. Change it to [1..4].

Also, during the generation do not allow a situation where some gate is
not attached to any plate at all. The random allocation is still the
same, but at the end attach all non-attached gates to one plate that has
the minimal number of attached gates.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>