Change definition of POS_CENTER_Y to be in the map center
Previously the whole height (together with the status at the bottom) was
taken into account, now only the map size is considered. This improves
most of the cases. Except for one case: currently "Initializing level…"
is shown slightly above the window center, but consistent with the rest
of the flash cases.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>