Add support for char/enemy power as alternative mode to health+attack
Now it is possible to configure either char.health or char.power. In the
first mode, the attack of char and enemy is used to calculate the mutual
damage and to update mutual health. There is no change to this mode.

In the new power mode, there is no health and attack, but instead power
is defined for char and for enemies. The rule is if the char has higher
power on encounter, it wins and absorbs the enemy power. There is no
continuous battle like in the health+attack mode, instead the win-lose is
determined in one turn and the winner gains the power of the opponent.

The 2 modes are distinguashed by presense of char.power when not None.

This new mode is intended for upcoming HeroPuzzle. But it may be useful
for other general battles. For it, numeric config key "char_power" should
be defined in the level config instead of "char_health".

Also, refactor a new function draw_actor_hint.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>