Add new drop class drop_might that is a power potion with factors
Now it is possible to instantiate new drop kind, it may look like a
potion or a fist (depending on the theme).

It affects the char power depending on drop's args:
	* op='×', factor  ->  power = power * factor
	* op='÷', factor  ->  power = power // factor
	* op='+', factor  ->  power = power + factor
	* op='-', factor  ->  power = power - factor

The new "might" sprite is created by us using AI and edited in GIMP.
May be freely distributed as Free Art.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>