Change enemy encounter mechanics in power mode
In power mode now the character does not stop to beat or kill enemy as in
health-attach mode. But either wins and moves forward or dies in place.

The kill-enemy process still has the same sound and falling animation,
but this is done in parallel with the character stepping into the enemy
as in the normal move.

With barrel in power mode, the enemy over barrel is killed
unconditionally. Otherwise this would stall the auto-solution in barrel
puzzle when the enemy on the way is stronger. Although the issue is
currently theoretical only, since no enemies are auto-created right now
in power mode.

Change the logic of who wins when the power of char and enemy is the
same. Make char win and double the power in this case.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>