Add new PuzzleInfo class to aid in automatic puzzle documentation
Provide an easy way to document each puzzle in info/puzzles/ directory.

The text format per puzzle is: the title (one line), the description
(several paragraphs), the goal (like: "Goal: Reach the finish.").

This functionality is supposed to be used in external tools that create
automatic documentation (including in the web site), as well as from the
game itself in the future.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>