Add filetools.py to help with some common file operations
Currently there is one helper class and one helper function.

Function get_dir_names lists sorted file names without extention given
the directory and the file extention.

Class RegenerateFile takes the target file and provides a write/print
interface to generate it. However it actually creates the new file with
".new" extention and then either moves it to the target file (and prints
"Regenerated") if the file content was changed, or prints "Not changed"
and removes the newly created identical file.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>