Add static html resources and simple template system
The idea is to have a single template for all html pages.

Every html page consists of: logo, sidebar, header and content.

Sidebar is different per category. There are currently 5 categories:
Home, Puzzles, Screenshots, Themes, Tutorial. They are all present in the
header as large horizontal buttons. They are also duplicated in sidebar
as smaller vertical buttons to suit narrowly sized browser windows.

There is a simple template system implemented in bash/perl. The rational
is not to depend on external systems in this game. The html is not
planned to be complex.

Some elements (for actual html pages and for category sidebars) are going
to be auto-generated from other project resources by dedicated scripts.

Some html pages are going to be auto-generated by dedicated scripts.

Signed-off-by: Mikhael Goikhman <migo@freeshell.org>